requirements.txt
flask==2.1.2
PyPDF2==2.5.0
nltk==3.7

# thumbnails
preview-generator==0.29

# pptx text extraction
python-pptx==0.6.21

# html text extraction
bs4==0.0.1

# docx text extraction
python-docx==0.8.11
